Day-0 config; logging; removed debug
Implemented Day-0 configuration. The python script now retrieves
the console ports for each node and uses Expect scripts to apply
the config. The reason Expect is required is that multiple vendors
need to be configured the same way.

Added outputs for each stage and disabled the output of Expect
scripts because they dumped a lot of output.

Removed debug mode, the final configuration is saved to the
topology_full.yml file after every run.

Also did some prettification.